content: generateTrackChronologyLinks: inline scoping logic
